+++ glade_forms/lib/src/core/input/glade_string_input.dart
@@ -1,4 +1,6 @@
+import 'package:glade_forms/src/core/error/glade_error_keys.dart';
 import 'package:glade_forms/src/core/input/glade_input.dart';
+import 'package:glade_forms/src/validator/part/satisfy_predicate_part.dart';
 import 'package:glade_forms/src/validator/specialized/string_validator.dart';
 
 /// A string input.
@@ -34,4 +36,13 @@ class GladeStringInput extends GladeInput<String> {
               ? (validator?.call(StringValidator()..notEmpty()) ?? (StringValidator()..notEmpty()).build())
               : (validator?.call(StringValidator()) ?? StringValidator().build()),
         );
+
+  /// Maximum length of the string if `maxLength()` standard validator is used.
+  int? getMaxLength({Object? key}) {
+    final validator = validatorInstance.tryFindValidatorPart(key ?? GladeErrorKeys.stringMaxLength);
+
+    if (validator == null) return null;
+
+    return validator is SatisfyPredicatePart ? (validator as SatisfyPredicatePart).metaData as int? : null;
+  }
 }