+++ glade_forms/lib/src/validator/validator_instance.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:glade_forms/src/core/core.dart';
 import 'package:glade_forms/src/validator/part/input_validator_part.dart';
 import 'package:glade_forms/src/validator/validator_error/glade_validator_error.dart';
@@ -8,6 +9,7 @@ class ValidatorInstance<T> {
   final bool stopOnFirstError;
 
   GladeInput<T>? _input;
+
   final List<InputValidatorPart<T>> _parts;
 
   ValidatorInstance({
@@ -37,5 +39,20 @@ class ValidatorInstance<T> {
     }
 
     return ValidatorResult(errors: errors, associatedInput: _input);
+  }
+
+  InputValidatorPart<T>? tryFindValidatorPart(Object key) {
+    return _parts.firstWhereOrNull((part) => part.key == key);
+  }
+
+  InputValidatorPart<T> findValidatorPart(Object key) {
+    return _parts.firstWhere(
+      (part) => part.key == key,
+      orElse: () => throw ArgumentError('No validator part found for key: $key'),
+    );
+  }
+
+  bool hasDeclaredValidator(Object key) {
+    return _parts.any((part) => part.key == key);
   }
 }